Match Java type 'long' to result of postgres count()
Postgres docs say that count() yields Postgres numerical type bigint [1]
which is 8 bytes according to [2].

This means that our Dao classes and pagination support should upgrade from
using 'int' to 'long' where applicable.

[1] https://www.postgresql.org/docs/8.2/functions-aggregate.html
[2] https://www.postgresql.org/docs/9.1/datatype-numeric.html

+++ textrepo-app/src/main/java/nl/knaw/huc/api/ResultPage.java
@@ -7,10 +7,10 @@ import java.util.List;
 public class ResultPage<T> {
 
   private final List<T> items;
-  private final int total;
+  private final long total;
   private final ResultPageParams page;
 
-  public ResultPage(List<T> items, int total, ResultPageParams page) {
+  public ResultPage(List<T> items, long total, ResultPageParams page) {
     this.items = items;
     this.total = total;
     this.page = page;
@@ -22,7 +22,7 @@ public class ResultPage<T> {
   }
 
   @JsonProperty
-  public int getTotal() {
+  public long getTotal() {
     return total;
   }
